changed to only int coordinates

+++ Point.py
@@ -8,12 +8,12 @@ class Point:
     def __init__(self, x, y, walls_on=None):
         if walls_on is None:
             walls_on = []
-        self.x: float = x
-        self.y: float = y
+        self.x: int = x
+        self.y: int = y
         self.walls_on: List[Line.Line] = walls_on
         self.seen_by: List[Point] = []
 
-    def get_coordinates(self) -> (float, float):
+    def get_coordinates(self) -> (int, int):
         return self.x, self.y
 
     def __repr__(self):
